adding basic api endpoints and basic tests.
this is one long commit.  sorry about that. :(  That was bad of me.

This uses expressjs to serve http requests for a backend api service.
(It is not a RESTful service.)

The service roughly emulates that from v2-cardano-graphql.js in the
yoroi-backend-service repository.

For each request, it is using axios to query a graphql server.

The haskeller in me whats to use Either types to capture errors, and
have a top-level server decide if the Either type should become a 500
error or some other exception.  Therefore I make use of a UtilEither
type defined in utils/index.js.

Apart from that, I think the code is straightforward.  Please read
the README.md before reviewing, however.