Feature/logging and error handling (#4)
* add - sending 500 errors when we have an error.

* fix - endpoint address for tests.

* add - middleware to log and handle errors.

* mod - throwing errors instead of trying to log and handle them.

I don't like functions that randomly throw errors.  I've tried (but haven't entirely succeeded
) to use UtilEither for functions that can error, but due to magic with async/await it is
possible for a graphql request to become in error in the ask* functions.

Anyhow - the top level methods are now throwing errors.  I guess something has got to throw
an error, or else I have to write a bunch of the same code over and over again.  so it
may as well be the top level handlers.

* fix - bestblock was not returning slotNo.

the query that I niavely copied was using slotDuration for slotNo.  silly.

* fix - tx history outputs should be ordered by tx index asc, inputs by input tx id asc.

* fix - filterUsed was not sending graphql params correctly.

Co-authored-by: Matthew Eric Bassett <matthew@emurgo.io>