Update tx history endpoint for Shelley txs [ch5527]
After about a half day of looking, I cannot find `ttl` anwhere in the
schema created by cardano-db-schema.  On Seba's advice, I have just used
the maxint.
`ttl` should be a slot_no, and these are stored in postgresql as a `uinteger`
which is just an `integer` ie `int4`  whose value >0.  So took the largest int4.

Certificates are not implemented.