add - test on shelley certificates.
these is a pretty lame test.
We cannot follow the previous test strategy of comparing to known good results because we don't have any
known good results.

Instead, we just check to ensure that the return objects are sensible.
We only focus on the two types of certs that have complicated subfields.

The addresses for the tests I just pulled out of mainnet because they had certs that I wanted to inspect.

The tests mostly ensure that the postgresql json encoding works correctly.