Feature/certs (#21)
* add - creating combined certificates view on startup.

The formal spec for the Cardano ledge
(https://hydra.iohk.io/build/3553697/download/1/ledger-spec.pdf)
specifies 7 certs, which correspond to the `formalType` column
in the view created.

Note that DCert_genesis is not implemented by cardano-db-sync, so it
does not appear here.

cardano-db-sync does not have a `certificate` table, instead it has a
table for each of the 6 certificate types that it implements.
The view created here does a `UNION ALL` so we can easily query and
aggregate them in, say, transaction history requests.

Resolving the certificates from cardano-db-sync is tedious, one should
refer to
cardano-db-sync:cardano-db-sync/src/Cardano/DbSync/Era/Shelley/Insert.hs
to resolve which certificates go into which table.

DCert_mir actually goes into 2 tables: `reserve` and `treasury`, which
suggests that the `cert_index` column might not be accurate - I haven't
read anything in the spec that suggests that a tx could not contain both
a `reserve` DCert_mir and a `treasury` DCert_mid.
The actual problem occurs with the `group by` clause in both of queries
for both tables - we need to get all the addresses for each DCert_mir,
which we do via a `group by` and `json_agg`.  But then we need to do
`cert_index`.  I assume that a tx has multiple Mirs with multiple addresses
that I should take the max `cert_index`. but if those are split between
`reserve` and `treasury` which max `cert_index` is correct?

Considering that `treasury` is empty at the time I authored and rebased this
commit, I think that it is OK for now.

* refactor - moving transaction history types to own file.

* add - adding typescript defs for cert objects.

* add - certificate field in transaction history endpoint

* add - test on shelley certificates.

these is a pretty lame test.
We cannot follow the previous test strategy of comparing to known good results because we don't have any
known good results.

Instead, we just check to ensure that the return objects are sensible.
We only focus on the two types of certs that have complicated subfields.

The addresses for the tests I just pulled out of mainnet because they had certs that I wanted to inspect.

The tests mostly ensure that the postgresql json encoding works correctly.

* mod - updating readme

* mod - adding certIndex to certificates in txHistory.

* mod - more changes to readme, specifying that addresses are coming in as hex.

Co-authored-by: Matthew Eric Bassett <matthew@emurgo.io>