mod - tx history endpoint accepts payment creds as well as addresses.
this requires commit e30563ec55b52ceeb6df48d89286a879000b0305 from cardano-db-sync
without this commit this code is not performant.

normally, you would think that we could do something like
	...
	WHERE ...
  	   OR encode(payment_cred,'hex') = ANY(($1)::varchar array)
	...
But that wouldn't work unless we put a function index on the tx_out.payment_cred
column, and I think that is unwise.
So we transform from string to bytea in js-land.
Now, bytea from psql become Buffer in js-land, so you would naively think that
a Buffer is what we need to send back.  But the naive thought is wrong.
Rather the pg library will convert a string it sees that starts with a \x
(as byteas appear in the console in psql).  So we must convert our strings thus.
But to avoid errors we will need to filter out any strings that aren't actually
a hex.
this is where the more complicated code comes from.